// kservices_ts_api/src/api/domain/entities/v3/sme.entity.ts
export type getRemarksRequest = {
  sme_id: number;
  call_direction: string;
  session_id: string;
  limit: number;
};

export type setRemarksRequest = {
  sme_id: number;
  call_direction: string;
  session_id: string;
  remarks: string;
  customer_number: string;
  created_by: number;
  start_date_time: string;
  updated_date_time: string;
};

export type fetchListRemarksRequest = {
  sme_id: number;
  call_direction: string;
  customer_number: number;
  limit: number;
};